requirements.txt
Flask~=2.1.3
numpy~=1.23.1
opencv-python~=4.6.0.66
pytesseract~=0.3.9
python-docx~=0.8.11
Werkzeug~=2.1.2
pdf2image~=1.16.0